Remove hard-coded paths
This deletes the local reference to the stash file so that these
functions can take advantage of the dynamic file path helpers that will
write standups to the user's home directory. Now they are ready to be
re-used for not just writing and reading the stash file but also
archiving standups as well.