Flush stdout
This ensures that standard output is not buffered (which it is by
default). Buffering causes the calls to `putStr*` print after the user
has already been prompted for input which is a confusing experience.

There are performance implications when turning off buffering entirely
as I'm doing here, however, they're negligible for such a small program.